+++ CHANGELOG.md
@@ -2,9 +2,14 @@
 
 ## Unreleased
 
+## Changed
+
+- Passing a limit without an offset will now set the offset to 0.
+- Passing a page size without a page will now set the page to 1.
+
 ## [0.6.1] - 2020-06-17
 
-## Changed
+### Changed
 
 - Add Flop to Meta struct.